requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
transformers
torch
pandas
huggingface_hub
